requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
app-store-scraper==0.3.5
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==3.0.1
defusedxml==0.7.1
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
dnspython==2.3.0
exceptiongroup==1.1.2
frozenlist==1.3.3
google-play-scraper==1.2.6
gunicorn==20.1.0
h11==0.14.0
idna==2.9
jira==3.4.1
Markdown==3.4.1
multidict==6.0.4
oauthlib==3.2.2
openai==0.27.2
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
PySocks==1.7.1
pytz==2022.7.1
requests==2.23.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
six==1.16.0
slack-bolt==1.16.2
slack-sdk==3.20.0
slackclient==2.9.4
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.2.4
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.25.6
wsproto==1.2.0
yarl==1.8.2
